moved utilities to utils.py
